fix: 커스텀 훅으로 변경
